Reorganize component file structure
