+++ client/app/app-routing.module.ts
@@ -4,13 +4,13 @@ import { RouterModule, Routes } from '@angular/router';
 import { OrganizationDashboardComponent } from './components/organization-dashboard/organization-dashboard.component';
 import { LoginComponent } from './components/login/login.component';
 import { Guard } from './guard';
-import { VolunteerCheckInComponent } from './components/volunteer-check-in/volunteer-check-in.component';
+import { LocationCheckInComponent } from './components/location-check-in/location-check-in.component';
 import { GettingStartedComponent } from './components/getting-started/getting-started.component';
 
 const routes: Routes = [
   { path: 'login', component: LoginComponent },
   { path: 'dashboard', component: OrganizationDashboardComponent, canActivate : [Guard] },
-  { path: 'checkin/:id', component: VolunteerCheckInComponent, canActivate : [Guard] },
+  { path: 'checkin/:id', component: LocationCheckInComponent, canActivate : [Guard] },
   { path: 'start', component: GettingStartedComponent },
   { path: '**', redirectTo: 'dashboard', pathMatch: 'full' }
 ];
