+++ client/app/components/location-check-in/check-in-form/check-in-form.component.spec.ts
