+++ client/app/components/location-check-in/check-in-form/signature-field/signature-field.component.html
