+++ client/app/components/location-check-in/location-check-in.component.html
