+++ client/app/components/location-check-in/location-check-in.component.spec.ts
@@ -3,17 +3,17 @@ import { MdTabsModule } from '@angular/material';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { MockComponent } from 'ng2-mock-component';
 
-import { VolunteerCheckInComponent } from './volunteer-check-in.component';
+import { LocationCheckInComponent } from './location-check-in.component';
 import { MockVolunteerService, VolunteerService } from '../../services/volunteer.service';
 
-describe('VolunteerCheckInComponent', () => {
-  let component: VolunteerCheckInComponent;
-  let fixture: ComponentFixture<VolunteerCheckInComponent>;
+describe('LocationCheckInComponent', () => {
+  let component: LocationCheckInComponent;
+  let fixture: ComponentFixture<LocationCheckInComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [
-        VolunteerCheckInComponent,
+        LocationCheckInComponent,
         MockComponent({ selector: 'app-check-in-form' }),
         MockComponent({ selector: 'app-new-volunteer-form', inputs: ['changeTab'] }),
         MockComponent({ selector: 'app-side-margins' })
@@ -29,7 +29,7 @@ describe('VolunteerCheckInComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(VolunteerCheckInComponent);
+    fixture = TestBed.createComponent(LocationCheckInComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
