+++ client/app/components/organization-dashboard/data-display/daily-hours-chart/daily-hours-chart.component.html
