+++ client/app/components/organization-dashboard/data-display/data-display.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-data-display',
+  templateUrl: './data-display.component.html',
+  styleUrls: ['./data-display.component.css']
+})
+export class DataDisplayComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
