+++ client/app/components/organization-dashboard/organization-dashboard.component.html
@@ -7,5 +7,5 @@
                       fxFlex.xs="4rem"
                       fxLayout="column">
   </app-volunteer-menu>
-  <app-visit-data-display fxFlex fxLayout="column"></app-visit-data-display>
+  <app-data-display fxFlex fxLayout="column"></app-data-display>
 </div>
