--- client/app/components/visit-data-display/visit-data-display.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-visit-data-display',
-  templateUrl: './visit-data-display.component.html',
-  styleUrls: ['./visit-data-display.component.css']
-})
-export class VisitDataDisplayComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
